reafactor: move goals formatter
